dropdown height aggressive calculation
The virutual repeat container was somewhat buggy with calculating the
actual height when less than the maxium dropdown items were remaining in
the search. Especially upon initial load with search text min-length 0.
This change reduces max dropdown height to the exact amount of choices
(min 1) if it is less than the maximum dropdown items.